hv: multi-arch construct barrier library
extract common barrier interface to include/lib/barrier.h,
and invoke the variant implementation of arch.

Tracked-On: #8803
Acked-by: Wang, Yu1 <yu1.wang@intel.com>
